Generalize Huawei to GridCharger
