+++ include/gridcharger/huawei/Provider.h
@@ -6,8 +6,16 @@
 #include <optional>
 #include <ArduinoJson.h>
 #include <TaskSchedulerDeclarations.h>
+#include <gridcharger/Provider.h>
 #include <gridcharger/huawei/HardwareInterface.h>
 #include <gridcharger/huawei/DataPoints.h>
+#include <gridcharger/huawei/Stats.h>
+#include <espMqttClient.h>
+#include <frozen/map.h>
+#include <frozen/string.h>
+#include <deque>
+#include <functional>
+#include <mutex>
 
 namespace GridChargers::Huawei {
 
@@ -17,19 +25,18 @@ namespace GridChargers::Huawei {
 #define HUAWEI_MODE_AUTO_EXT 2
 #define HUAWEI_MODE_AUTO_INT 3
 
-class Controller {
+class Provider : public ::GridChargers::Provider {
 public:
-    void init(Scheduler& scheduler);
-    void updateSettings();
-    void setFan(bool online, bool fullSpeed);
+    bool init() final;
+    void deinit() final;
+    void loop() final;
+
+    std::shared_ptr<::GridChargers::Stats> getStats() const final { return _stats; }
+    bool getAutoPowerStatus() const final { return _autoPowerEnabled; };
+
     void setProduction(bool enable);
     void setParameter(float val, HardwareInterface::Setting setting);
-    void setMode(uint8_t mode);
 
-    DataPointContainer const& getDataPoints() const { return _dataPoints; }
-    void getJsonData(JsonVariant& root) const;
-
-    bool getAutoPowerStatus() const { return _autoPowerEnabled; };
     uint8_t getMode() const { return _mode; };
 
     // determined through trial and error (voltage limits, R4850G2)
@@ -46,9 +53,11 @@ public:
     static constexpr float MAX_INPUT_CURRENT_LIMIT = 40.0f;
 
 private:
-    void loop();
     void _setParameter(float val, HardwareInterface::Setting setting, bool pollFeedback = false);
     void _setProduction(bool enable);
+
+    void setFan(bool online, bool fullSpeed);
+    void setMode(uint8_t mode);
 
     // these control the pin named "power", which in turn is supposed to control
     // a relay (or similar) to enable or disable the PSU using it's slot detect
@@ -56,40 +65,6 @@ private:
     void enableOutput();
     void disableOutput();
     gpio_num_t _huaweiPower;
-
-    template<DataPointLabel L>
-    void addValueInSection(JsonVariant& root,
-        std::string const& section, std::string const& name) const
-    {
-        auto oVal = _dataPoints.get<L>();
-        if (!oVal) { return; }
-
-        auto jsonValue = root["values"][section][name];
-        jsonValue["v"] = *oVal;
-        jsonValue["u"] = DataPointLabelTraits<L>::unit;
-        jsonValue["d"] = 2;
-    }
-
-    template<DataPointLabel L>
-    void addStringInSection(JsonVariant& root,
-        std::string const& section, std::string const& name) const
-    {
-        auto oVal = _dataPoints.get<L>();
-        if (!oVal) { return; }
-
-        auto jsonValue = root["values"][section][name];
-        jsonValue["value"] = *oVal;
-        jsonValue["translate"] = false;
-    }
-
-    void addStringInSection(JsonVariant& root,
-        std::string const& section, std::string const& name,
-        std::string const& value) const
-    {
-        auto jsonValue = root["values"][section][name];
-        jsonValue["value"] = value;
-        jsonValue["translate"] = true;
-    }
 
     Task _loopTask;
     std::unique_ptr<HardwareInterface> _upHardwareInterface;
@@ -99,6 +74,7 @@ private:
     uint8_t _mode = HUAWEI_MODE_AUTO_EXT;
 
     DataPointContainer _dataPoints;
+    std::shared_ptr<Stats> _stats = std::make_shared<Stats>();
 
     uint32_t _outputCurrentOnSinceMillis;         // Timestamp since when the PSU was idle at zero amps
     uint32_t _nextAutoModePeriodicIntMillis;      // When to set the next output voltage in automatic mode
@@ -108,8 +84,44 @@ private:
     uint8_t _autoPowerEnabledCounter = 0;
     bool _autoPowerEnabled = false;
     bool _batteryEmergencyCharging = false;
+
+    enum class Topic : unsigned {
+        LimitOnlineVoltage,
+        LimitOnlineCurrent,
+        LimitOfflineVoltage,
+        LimitOfflineCurrent,
+        LimitInputCurrent,
+        Mode,
+        Production,
+        FanOnlineFullSpeed,
+        FanOfflineFullSpeed
+    };
+
+    void subscribeTopics();
+    void unsubscribeTopics();
+
+    static constexpr frozen::string _cmdtopic = "huawei/cmd/";
+    static constexpr frozen::map<frozen::string, Topic, 9> _subscriptions = {
+        { "limit_online_voltage",   Topic::LimitOnlineVoltage },
+        { "limit_online_current",   Topic::LimitOnlineCurrent },
+        { "limit_offline_voltage",  Topic::LimitOfflineVoltage },
+        { "limit_offline_current",  Topic::LimitOfflineCurrent },
+        { "limit_input_current",    Topic::LimitInputCurrent },
+        { "mode",                   Topic::Mode },
+        { "production",             Topic::Production },
+        { "fan_online_full_speed",  Topic::FanOnlineFullSpeed },
+        { "fan_offline_full_speed", Topic::FanOfflineFullSpeed },
+    };
+
+    void onMqttMessage(Topic enumTopic,
+            const espMqttClientTypes::MessageProperties& properties,
+            const char* topic, const uint8_t* payload, size_t len);
+
+    // MQTT callbacks to process updates on subscribed topics are executed in
+    // the MQTT thread's context. we use this queue to switch processing the
+    // user requests into the main loop's context (TaskScheduler context).
+    std::deque<std::function<void()>> _mqttCallbacks;
 };
 
-} // namespace GridChargers::Huawei
 
-extern GridChargers::Huawei::Controller HuaweiCan;
+} // namespace GridChargers::Huawei
