+++ src/WebApi_mqtt.cpp
@@ -8,7 +8,6 @@
 #include "MqttHandleHass.h"
 #include "MqttHandlePowerLimiterHass.h"
 #include "MqttHandleInverter.h"
-#include "MqttHandleHuawei.h"
 #include "MqttHandlePowerLimiter.h"
 #include "MqttSettings.h"
 #include "WebApi.h"
@@ -18,6 +17,7 @@
 #include <powermeter/Controller.h>
 #include <AsyncJson.h>
 #include <solarcharger/Controller.h>
+#include <gridcharger/Controller.h>
 
 void WebApiMqttClass::init(AsyncWebServer& server, Scheduler& scheduler)
 {
@@ -316,13 +316,11 @@ void WebApiMqttClass::onMqttAdminPost(AsyncWebServerRequest* request)
         // Check if base topic was changed
         if (strcmp(config.Mqtt.Topic, root["mqtt_topic"].as<String>().c_str())) {
             MqttHandleInverter.unsubscribeTopics();
-            MqttHandleHuawei.unsubscribeTopics();
             MqttHandlePowerLimiter.unsubscribeTopics();
 
             strlcpy(config.Mqtt.Topic, root["mqtt_topic"].as<String>().c_str(), sizeof(config.Mqtt.Topic));
 
             MqttHandleInverter.subscribeTopics();
-            MqttHandleHuawei.subscribeTopics();
             MqttHandlePowerLimiter.subscribeTopics();
         }
     }
@@ -337,7 +335,7 @@ void WebApiMqttClass::onMqttAdminPost(AsyncWebServerRequest* request)
     MqttHandlePowerLimiterHass.forceUpdate();
     Battery.updateSettings();
 
-    MqttHandleHuawei.forceUpdate();
+    GridCharger.updateSettings();
     MqttHandlePowerLimiter.forceUpdate();
 
     SolarCharger.updateSettings();
